fix(cdn): update URL handling for CDN endpoints
