+++ project/backend/apps/aitoearn-server/src/common/utils/file.util.ts
@@ -1,5 +1,36 @@
 import * as fs from 'node:fs'
 import path from 'node:path'
+import { buildUrl, zodBuildUrl, zodTrimHost } from '@yikart/aws-s3'
+import { config } from '../../config'
+
+class FileUtil {
+  private hostUrl = config.awsS3.cdnEndpoint || config.awsS3.endpoint
+  public buildUrl(path = '') {
+    if (!path)
+      return path
+    return buildUrl(this.hostUrl, path)
+  }
+
+  zodBuildUrl() {
+    return zodBuildUrl(this.hostUrl)
+  }
+
+  /**
+   * 去除host部分，保留path部分
+   * @param url
+   * @returns
+   */
+  public trimHost(url: string) {
+    if (!url)
+      return url
+    return url.replace(this.hostUrl, '')
+  }
+
+  zodTrimHost() {
+    return zodTrimHost(this.hostUrl)
+  }
+}
+export const fileUtil = new FileUtil()
 
 enum Type {
   IMAGE = '图片',
