+++ project/backend/apps/aitoearn-server/src/content/mediaGroup.controller.ts
@@ -18,7 +18,7 @@ import { ApiTags } from '@nestjs/swagger'
 import { GetToken, TokenInfo } from '@yikart/aitoearn-auth'
 import { ApiDoc, AppException, ResponseCode, TableDto } from '@yikart/common'
 import { MediaGroup } from '@yikart/mongodb'
-import { fileUtil } from '../util/file.util'
+import { fileUtil } from '../common/utils/file.util'
 import {
   CreateMediaGroupDto,
   MediaGroupFilterDto,
