--- project/backend/apps/aitoearn-server/src/util/file.util.ts
@@ -1,31 +1,0 @@
-import { buildUrl, zodBuildUrl, zodTrimHost } from '@yikart/aws-s3'
-import { config } from '../config'
-
-class FileUtil {
-  private hostUrl = config.awsS3.endpoint
-  public buildUrl(path = '') {
-    if (!path)
-      return path
-    return buildUrl(this.hostUrl, path)
-  }
-
-  zodBuildUrl() {
-    return zodBuildUrl(this.hostUrl)
-  }
-
-  /**
-   * 去除host部分，保留path部分
-   * @param url
-   * @returns
-   */
-  public trimHost(url: string) {
-    if (!url)
-      return url
-    return url.replace(this.hostUrl, '')
-  }
-
-  zodTrimHost() {
-    return zodTrimHost(this.hostUrl)
-  }
-}
-export const fileUtil = new FileUtil()
